warnings: session.autocommit, subtransactions
